+++ src/Components/item.properties/ItemProperties.jsx
@@ -3,9 +3,9 @@ import {decodeIFCString} from '@bldrs-ai/ifclib'
 import Switch from '@mui/material/Switch'
 import Typography from '@mui/material/Typography'
 import {makeStyles, useTheme} from '@mui/styles'
-import useStore from '../store/useStore'
-import {createPropertyTable} from '../utils/itemProperties'
-import ExpansionPanel from './ExpansionPanel'
+import useStore from '../../store/useStore'
+import {createPropertyTable} from '../../utils/itemProperties'
+import ExpansionPanel from '../ExpansionPanel'
 
 
 /**
