+++ src/Components/side.drawer/SideDrawer.jsx
@@ -3,11 +3,11 @@ import {useLocation} from 'react-router-dom'
 import Drawer from '@mui/material/Drawer'
 import {Box} from '@mui/system'
 import {makeStyles, useTheme} from '@mui/styles'
-import useStore from '../store/useStore'
-import {getHashParams} from '../utils/location'
-import {preprocessMediaQuery} from '../utils/mediaQuery'
-import MobileDrawer from './MobileDrawer'
-import {MOBILE_WIDTH, useIsMobile} from './Hooks'
+import useStore from '../../store/useStore'
+import {getHashParams} from '../../utils/location'
+import {preprocessMediaQuery} from '../../utils/mediaQuery'
+import MobileDrawer from '../MobileDrawer'
+import {MOBILE_WIDTH, useIsMobile} from '../Hooks'
 import {PropertiesPanel, NotesPanel} from './SideDrawerPanels'
