+++ src/Components/side.drawer/SideDrawer.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import {act, render, renderHook} from '@testing-library/react'
-import useStore from '../store/useStore'
-import ShareMock from '../ShareMock'
+import useStore from '../../store/useStore'
+import ShareMock from '../../ShareMock'
 import SideDrawerWrapper from './SideDrawer'
