+++ src/Components/side.drawer/SideDrawerPanels.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import {act, render, renderHook} from '@testing-library/react'
-import ShareMock from '../ShareMock'
-import useStore from '../store/useStore'
+import ShareMock from '../../ShareMock'
+import useStore from '../../store/useStore'
 import {NotesPanel, PropertiesPanel} from './SideDrawerPanels'
